requirements.txt
numpy>=1.24
matplotlib>=3.7
networkx>=3.1
gymnasium>=0.29
stable-baselines3>=2.2.1
torch>=2.1
pyyaml>=6.0
tqdm>=4.66
